refactor: 스타일 구조 변경
- 개별 스타일 변수 대신 도메인 별 style object로 관리
- styles 디렉토리에 스타일 정의 파일 분리
