+++ src/app/(auth)/selectChannel/page.tsx
@@ -1,6 +1,6 @@
-import { pageTitleStyle } from '@/app/common.css';
 import CustomLink from '@/components/customLink/component';
 import MyChannelListFetcher from '@/fetcher/myChannelListFetcher/component';
+import { commonStyle } from '@/styles/common.css';
 import { Metadata } from 'next';
 
 export const metadata: Metadata = { title: '채널 선택' };
@@ -8,7 +8,7 @@ export const metadata: Metadata = { title: '채널 선택' };
 export default function SelectChannelPage() {
   return (
     <main>
-      <h1 className={pageTitleStyle}>채널 선택</h1>
+      <h1 className={commonStyle.pageTitle}>채널 선택</h1>
       <MyChannelListFetcher />
       <CustomLink href="/createChannel">
         새로운 채널 생성
