+++ src/app/(main)/channel/[channelId]/page.tsx
@@ -3,10 +3,10 @@ import ChannelProfileFetcher from '@/fetcher/channelProfileFetcher/component';
 import ChannelProfileSkeleton from '@/fetcher/channelProfileFetcher/skeleton';
 import ChannelVideoListFetcher from '@/fetcher/channelVideoListFetcher/component';
 import ChannelVideoListSkeleton from '@/fetcher/channelVideoListFetcher/skeleton';
+import { videoStyle } from '@/styles/video.css';
 import { fetchHandlerWithServerComponent } from '@/utils/handler';
 import { Metadata } from 'next';
 import { Suspense } from 'react';
-import { videoSectionStyle, videoSectionTitleStyle } from '@/app/common.css';
 
 interface ChannelPageProps { params: Promise<{ channelId: string }> }
 
@@ -30,8 +30,8 @@ export default async function ChannelPage({ params }: ChannelPageProps) {
       <Suspense fallback={<ChannelProfileSkeleton />}>
         <ChannelProfileFetcher channelId={channelId} />
       </Suspense>
-      <section className={videoSectionStyle}>
-        <h2 className={videoSectionTitleStyle}>업로드한 동영상</h2>
+      <section className={videoStyle.listSection}>
+        <h2 className={videoStyle.listSectionTitle}>업로드한 동영상</h2>
         <Suspense fallback={<ChannelVideoListSkeleton />}>
           <ChannelVideoListFetcher channelId={channelId} />
         </Suspense>
