--- src/app/(main)/studio/contents/[videoId]/page.module.css
@@ -1,5 +1,0 @@
-.title {
-  padding: 1.5rem 2rem;
-  font-weight: 700;
-  font-size: 1.75rem;
-}
