+++ src/app/(main)/studio/contents/page.tsx
@@ -1,10 +1,10 @@
 import StudioContentsFetcher from '@/fetcher/studioContentsFetcher/component';
-import { pageTitleStyle } from '@/app/common.css';
+import { commonStyle } from '@/styles/common.css';
 
 export default function StudioContentsPage() {
   return (
     <main>
-      <h1 className={pageTitleStyle}>내 콘텐츠 목록</h1>
+      <h1 className={commonStyle.pageTitle}>내 콘텐츠 목록</h1>
       <StudioContentsFetcher />
     </main>
   );
