+++ src/app/(main)/studio/info/page.tsx
@@ -1,10 +1,10 @@
-import { pageTitleStyle } from '@/app/common.css';
 import StudioInfoFetcher from '@/fetcher/studioInfoFetcher/component';
+import { commonStyle } from '@/styles/common.css';
 
 export default function StudioInfoPage() {
   return (
     <main>
-      <h1 className={pageTitleStyle}>내 채널 정보</h1>
+      <h1 className={commonStyle.pageTitle}>내 채널 정보</h1>
       <StudioInfoFetcher />
     </main>
   );
