+++ src/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
 import localFont from 'next/font/local';
-import './globals.css';
+import '@/styles/globals.css';
 import ToastProvider from '@/contexts/toast';
 import TokenProvider from '@/contexts/token';
 import { getAccessTokenCookie } from '@/serverActions/token';
